+++ src/jsfuncs.ts
@@ -13,6 +13,7 @@ import * as DataCtx from "./datactx";
 import type {DataEnvironment} from "./state";
 
 let DefaultJSFuncs : Record<string, JSFuncType> = {};
+let MAX_ITERS = 10000;
 
 function jsLen(v : HibikiVal) : number {
     if (v == null) {
@@ -28,12 +29,34 @@ function jsLen(v : HibikiVal) : number {
     return 1;
 }
 
-function jsIndexOf(str : string, ...rest : any[]) {
-    if (str == null) {
+function jsIndexOf(params : DataCtx.HibikiParamsObj) {
+    let data = params.getArg(null, 0);
+    if (data == null) {
         return -1;
     }
-    // @ts-ignore
-    return str.indexOf(...rest);
+    let fromIndex = params.getArg("fromindex", 2);
+    if (fromIndex == null) {
+        fromIndex = 0;
+    }
+    if (typeof(fromIndex) !== "number") {
+        throw new Error(sprintf("fn:indexof, fromindex (3rd argument) must be a number, got type=%s", DataCtx.hibikiTypeOf(fromIndex)));
+    }
+    if (typeof(data) === "string") {
+        let findArg = params.getArg(null, 1);
+        if (findArg == null) {
+            return 0;
+        }
+        if (typeof(findArg) !== "string") {
+            throw new Error(sprintf("fn:indexof[string], 2nd argument must be string, got type=%s", DataCtx.hibikiTypeOf(findArg)));
+        }
+        return data.indexOf(findArg, fromIndex);
+    }
+    let [arrObj, isArr] = DataCtx.asArray(data, false);
+    if (!isArr) {
+        throw new Error(sprintf("fn:indexof, 1st argument must be string or array, got type=%s", DataCtx.hibikiTypeOf(data)));
+    }
+    let findArg = params.getArg(null, 1);
+    return arrObj.indexOf(findArg, fromIndex);
 }
 
 function jsMin(...rest : HibikiVal[]) : number {
@@ -123,24 +146,285 @@ function jsSlice(params : DataCtx.HibikiParamsObj) : HibikiVal[] {
     return slicedIndexArr.map((idx) => rawLv.subArrayIndex(idx));
 }
 
-function jsPush(arr : HibikiVal[], ...rest : HibikiVal[]) {
-    if (arr == null || !mobx.isArrayLike(arr)) {
-        return [...rest];
+function jsPush(val : HibikiVal, ...rest : HibikiVal[]) {
+    let [arrObj, isArr] = DataCtx.asArray(val, true);
+    if (!isArr) {
+        return null;
     }
-    let rtn = [...arr, ...rest];
+    if (arrObj == null) {
+        arrObj = [];
+    }
+    let rtn = [...arrObj, ...rest];
     return rtn;
 }
 
-function jsPop(arr : HibikiVal[], ...rest : HibikiVal[]) {
-    if (arr == null) {
-        return [];
+function jsUnshift(val : HibikiVal, ...rest : HibikiVal[]) {
+    let [arrObj, isArr] = DataCtx.asArray(val, true);
+    if (!isArr) {
+        return null;
     }
-    else if (!mobx.isArrayLike(arr)) {
-        return [];
+    if (arrObj == null) {
+        arrObj = [];
     }
-    let rtn = [...arr];
-    rtn.pop();
+    let rtn = [...rest, ...arrObj];
     return rtn;
+}
+
+function jsPop(val : HibikiVal, num : HibikiVal) : HibikiVal[] {
+    if (val == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(val, false);
+    if (!isArr) {
+        return null;
+    }
+    let rtn = [...arrObj];
+    if (typeof(num) === "number" && !isNaN(num) && num >= 0 && num <= MAX_ITERS) {
+        for (let i=0; i<num; i++) {
+            rtn.pop();
+        }
+    }
+    else {
+        rtn.pop();
+    }
+    return rtn;
+}
+
+function jsShift(val : HibikiVal, num : HibikiVal) {
+    if (val == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(val, false);
+    if (!isArr) {
+        return null;
+    }
+    let rtn = [...arrObj];
+    if (typeof(num) === "number" && !isNaN(num) && num >= 0 && num <= MAX_ITERS) {
+        for (let i=0; i<num; i++) {
+            rtn.shift();
+        }
+    }
+    else {
+        rtn.shift();
+    }
+    return rtn;
+}
+
+function jsJoin(val : HibikiVal, str : HibikiVal) : string {
+    if (val == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(val, false);
+    if (!isArr) {
+        return null;
+    }
+    if (str == null) {
+        str = "";
+    }
+    if (typeof(str) !== "string") {
+        throw new Error(sprintf("fn:join 2nd argument must be type string, got type=%s", DataCtx.hibikiTypeOf(str)));
+    }
+    return arrObj.join(str);
+}
+
+function jsArrFilter(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : HibikiVal[] {
+    let arrArg = params.getArg("arr", 0);
+    let lambdaArg = params.getArg("expr", 1);
+    if (arrArg == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(arrArg, false);
+    if (!isArr) {
+        return null;
+    }
+    if (lambdaArg == null) {
+        return arrObj;
+    }
+    if (!(lambdaArg instanceof DataCtx.LambdaValue)) {
+        throw new Error(sprintf("fn:filter 2nd argument must be type lambda, got type=%s", DataCtx.hibikiTypeOf(lambdaArg)));
+    }
+    let lambda : DataCtx.LambdaValue = lambdaArg;
+    let rtn : HibikiVal[] = [];
+    for (let idx=0; idx<arrObj.length; idx++) {
+        let params = {elem: arrObj[idx], index: idx};
+        let include = lambda.invoke(dataenv, new DataCtx.HibikiParamsObj(params));
+        if (include) {
+            rtn.push(arrObj[idx]);
+        }
+    }
+    return rtn;
+}
+
+function jsArrSome(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : boolean {
+    let findIndex = jsArrFindIndex(params, dataenv);
+    return (findIndex !== -1);
+}
+
+function jsArrEvery(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : boolean {
+    let arrArg = params.getArg("arr", 0);
+    let lambdaArg = params.getArg("expr", 1);
+    if (arrArg == null) {
+        return true;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(arrArg, false);
+    if (!isArr) {
+        return false;
+    }
+    if (lambdaArg == null) {
+        return true;
+    }
+    if (!(lambdaArg instanceof DataCtx.LambdaValue)) {
+        throw new Error(sprintf("fn:every 2nd argument must be type lambda, got type=%s", DataCtx.hibikiTypeOf(lambdaArg)));
+    }
+    let lambda : DataCtx.LambdaValue = lambdaArg;
+    let rtn : HibikiVal[] = [];
+    for (let idx=0; idx<arrObj.length; idx++) {
+        let params = {elem: arrObj[idx], index: idx};
+        let include = lambda.invoke(dataenv, new DataCtx.HibikiParamsObj(params));
+        if (!include) {
+            return false;
+        }
+    }
+    return true;
+}
+
+function jsArrMap(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : HibikiVal[] {
+    let arrArg = params.getArg("arr", 0);
+    let lambdaArg = params.getArg("expr", 1);
+    if (arrArg == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(arrArg, false);
+    if (!isArr) {
+        return null;
+    }
+    if (lambdaArg == null) {
+        return arrObj;
+    }
+    if (!(lambdaArg instanceof DataCtx.LambdaValue)) {
+        throw new Error(sprintf("fn:map 2nd argument must be type lambda, got type=%s", DataCtx.hibikiTypeOf(lambdaArg)));
+    }
+    let lambda : DataCtx.LambdaValue = lambdaArg;
+    let rtn : HibikiVal[] = [];
+    for (let idx=0; idx<arrObj.length; idx++) {
+        let params = {elem: arrObj[idx], index: idx};
+        let newVal = lambda.invoke(dataenv, new DataCtx.HibikiParamsObj(params));
+        rtn.push(newVal);
+    }
+    return rtn;
+}
+
+function jsArrConcat(...rest : HibikiVal[]) : HibikiVal[] {
+    let rtn = [];
+    if (rest == null || rest.length === 0) {
+        return rtn;
+    }
+    for (let i=0; i<rest.length; i++) {
+        if (rest[i] == null) {
+            continue;
+        }
+        let [arrObj, isArr] = DataCtx.asArray(rest[i], false);
+        if (!isArr) {
+            throw new Error(sprintf("fn:concat, argument #%d is not an array, got type=%s", i, DataCtx.hibikiTypeOf(rest[i])));
+        }
+        if (arrObj.length === 0) {
+            continue;
+        }
+        rtn = rtn.concat(arrObj);
+    }
+    return rtn;
+}
+
+function jsArrFindIndexInternal(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment, fnName : string) : [HibikiVal, number] {
+    let arrArg = params.getArg("arr", 0);
+    let lambdaArg = params.getArg("expr", 1);
+    let fromIndex = params.getArg("fromindex");
+    if (fromIndex == null) {
+        fromIndex = 0;
+    }
+    if (typeof(fromIndex) !== "number") {
+        throw new Error(sprintf("%s 'fromindex' argument must be type number, got type=%s", fnName, DataCtx.hibikiTypeOf(fromIndex)));
+    }
+    if (arrArg == null) {
+        return [null, -1];
+    }
+    let [arrObj, isArr] = DataCtx.asArray(arrArg, false);
+    if (!isArr) {
+        return [null, -1];
+    }
+    if (lambdaArg == null) {
+        return [null, -1];
+    }
+    if (!(lambdaArg instanceof DataCtx.LambdaValue)) {
+        throw new Error(sprintf("%s 2nd argument must be type lambda, got type=%s", fnName, DataCtx.hibikiTypeOf(lambdaArg)));
+    }
+    let lambda : DataCtx.LambdaValue = lambdaArg;
+    for (let idx=fromIndex; idx<arrObj.length; idx++) {
+        let params = {elem: arrObj[idx], index: idx};
+        let include = lambda.invoke(dataenv, new DataCtx.HibikiParamsObj(params));
+        if (include) {
+            return [arrObj[idx], idx];
+        }
+    }
+    return [null, -1];
+}
+
+function jsArrFindIndex(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : number {
+    let [rtn, idx] = jsArrFindIndexInternal(params, dataenv, "find:findindex");
+    return idx;
+}
+
+function jsArrFind(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : HibikiVal {
+    let [rtn, idx] = jsArrFindIndexInternal(params, dataenv, "fn:find");
+    return rtn;
+}
+
+function jsArrReduce(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : HibikiVal {
+    let arrArg = params.getArg("arr", 0);
+    let lambdaArg = params.getArg("expr", 1);
+    let initialValue = params.getArg("initialvalue", 2);
+    let hasInitialValue = params.hasArg("initialvalue", 2);
+    if (arrArg == null) {
+        return null;
+    }
+    let [arrObj, isArr] = DataCtx.asArray(arrArg, false);
+    if (!isArr) {
+        return null;
+    }
+    if (lambdaArg == null) {
+        return arrObj;
+    }
+    if (!(lambdaArg instanceof DataCtx.LambdaValue)) {
+        throw new Error(sprintf("fn:reduce 2nd argument must be type lambda, got type=%s", DataCtx.hibikiTypeOf(lambdaArg)));
+    }
+    let lambda : DataCtx.LambdaValue = lambdaArg;
+    let value : HibikiVal = (hasInitialValue ? initialValue : null);
+    for (let idx=(hasInitialValue ? 1 : 0); idx<arrObj.length; idx++) {
+        let params = {elem: arrObj[idx], index: idx, value: value};
+        value = lambda.invoke(dataenv, new DataCtx.HibikiParamsObj(params));
+    }
+    return value;
+}
+
+function jsArrReverse(params : DataCtx.HibikiParamsObj, dataenv : DataEnvironment) : HibikiVal[] {
+    let rawVal = params.getArg("arr", 0);
+    if (rawVal == null) {
+        return null;
+    }
+    let val = DataCtx.resolveLValue(rawVal);
+    let [arrObj, isArr] = DataCtx.asArray(val, false);
+    if (!isArr) {
+        return null;
+    }
+    let makeRefs = DataCtx.valToBool(params.getArg("makerefs"));
+    if (makeRefs && !(rawVal instanceof DataCtx.LValue)) {
+        console.log("WARNING fn:reverse makerefs=true, but data is not a reference");
+    }
+    if (!makeRefs || !(rawVal instanceof DataCtx.LValue)) {
+        return arrObj.reverse();
+    }
+    let rawLv = rawVal as DataCtx.LValue;
+    return arrObj.map((_, idx) => rawLv.subArrayIndex(idx)).reverse();
 }
 
 function jsSetAdd(arr : HibikiVal[], ...rest : HibikiVal[]) {
@@ -375,6 +659,7 @@ function jsReplaceAll(str : HibikiVal, findStr : HibikiVal, replaceStr : HibikiV
     if (typeof(replaceStr) !== "string") {
         replaceStr = DataCtx.valToString(replaceStr);
     }
+    // @ts-ignore - replaceAll exists
     return str.replaceAll(findStr, replaceStr);
 }
 
@@ -652,6 +937,21 @@ reg("splice", jsSplice, true);
 regParamFn("slice", jsSlice);
 reg("push", jsPush, true);
 reg("pop", jsPop, true);
+reg("unshift", jsUnshift, true);
+reg("shift", jsShift, true);
+reg("join", jsJoin, true);
+regParamFn("filter", jsArrFilter);
+regParamFn("map", jsArrMap);
+regParamFn("find", jsArrFind);
+regParamFn("findindex", jsArrFindIndex);
+regParamFn("reduce", jsArrReduce);
+regParamFn("reverse", jsArrReverse);
+regParamFn("every", jsArrEvery);
+regParamFn("some", jsArrSome);
+reg("concat", jsArrConcat, true);
+
+// concat
+
 
 // obj functions
 reg("objkeys", jsObjKeys, true);
@@ -660,7 +960,7 @@ reg("objallkeys", jsObjAllKeys, true);
 
 // string functions
 reg("substr", jsSubstr, true);
-reg("indexof", jsIndexOf, false);
+regParamFn("indexof", jsIndexOf);  // works for strings or arrays
 reg("uppercase", jsUpperCase, true);
 reg("lowercase", jsLowerCase, true);
 reg("sprintf", jsSprintf, true);